+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Remove deprecated Raspberry Pi Enable Camera option
  - Fix no image when libcamera AWB turned off
  - Fix Rsync Backup Function ([#1314](https://github.com/kizniche/Mycodo/issues/1314))
+ - Fix Activate/Deactivate Controller Widget not recognizing Custom Functions
 
 ### Features